fix(editor): address LSP CI regressions
