+++ tests/tooling/editor-integrations.test.ts
@@ -217,54 +217,6 @@ test("vscode-vize wires art-vue documents into editor features", () => {
   assert.match(extensionSource, /Trace\.(Verbose|Messages|Off)/);
 });
 
-test("editor adapters consistently route Vue documents to the Vize LSP", () => {
-  const vscodeManifest = readJson<{
-    activationEvents?: string[];
-    contributes?: {
-      typescriptServerPlugins?: unknown;
-    };
-  }>("editors/vscode/package.json");
-  for (const language of ["vue", "art-vue", "html"]) {
-    assert.ok(
-      vscodeManifest.activationEvents?.includes(`onLanguage:${language}`),
-      `missing VS Code activation for ${language}`,
-    );
-  }
-  assert.equal(vscodeManifest.contributes?.typescriptServerPlugins, undefined);
-
-  const nvimConfig = fs.readFileSync(path.join(root, "editors/nvim/lua/vize/config.lua"), "utf-8");
-  assert.match(nvimConfig, /cmd = \{ "vize", "lsp" \}/);
-  assert.match(nvimConfig, /filetypes = \{ "vue", "art-vue" \}/);
-  assert.match(nvimConfig, /editor = true/);
-  assert.match(nvimConfig, /lint = true/);
-
-  const vimConfig = fs.readFileSync(path.join(root, "editors/vim/autoload/vize.vim"), "utf-8");
-  assert.match(vimConfig, /'cmd': \['vize', 'lsp'\]/);
-  assert.match(vimConfig, /'allowlist': \['vue', 'art-vue'\]/);
-  assert.match(vimConfig, /'editor': v:true/);
-  assert.match(vimConfig, /'lint': v:true/);
-
-  const helixConfig = fs.readFileSync(path.join(root, "editors/helix/languages.toml"), "utf-8");
-  assert.match(helixConfig, /\[language-server\.vize\][\s\S]*command = "vize"/);
-  assert.match(helixConfig, /\[language-server\.vize\][\s\S]*args = \["lsp"\]/);
-  assert.match(helixConfig, /name = "vue"[\s\S]*language-servers = \["vize"\]/);
-  assert.match(helixConfig, /name = "art-vue"[\s\S]*language-id = "art-vue"/);
-
-  const emacsConfig = fs.readFileSync(path.join(root, "editors/emacs/vize.el"), "utf-8");
-  assert.match(emacsConfig, /defcustom vize-eglot-command '\("vize" "lsp"\)/);
-  assert.match(emacsConfig, /vue-mode vue-ts-mode web-mode vize-vue-mode vize-art-vue-mode/);
-  assert.match(emacsConfig, /recommended \. \(:editor t :ecosystem t :lint t :typecheck t\)/);
-
-  const zedManifest = fs.readFileSync(path.join(root, "editors/zed/extension.toml"), "utf-8");
-  assert.match(zedManifest, /\[language_servers\.vize\][\s\S]*languages = \["Vue", "Art Vue"\]/);
-  assert.match(zedManifest, /"Vue" = "vue"/);
-  assert.match(zedManifest, /"Art Vue" = "art-vue"/);
-
-  const zedSource = fs.readFileSync(path.join(root, "editors/zed/src/lib.rs"), "utf-8");
-  assert.match(zedSource, /const SERVER_BINARY: &'static str = "vize"/);
-  assert.match(zedSource, /unwrap_or_else\(\|\| vec!\["lsp"\.to_string\(\)\]\)/);
-});
-
 test("vscode-vize grammar keeps quote-aware block lookaheads", () => {
   type GrammarCapture = {
     name?: string;
